bugfix : Id 타입 변경에 따른 수정 및 기타 작업

+++ src/main/java/pinup/backend/report/query/controller/ReportController.java
@@ -24,7 +24,7 @@ public class ReportController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<ReportSpecificResponse> getReportById(@PathVariable Integer id) {
+    public ResponseEntity<ReportSpecificResponse> getReportById(@PathVariable Long id) {
         return ResponseEntity.ok(reportService.getSpecificReport(id));
     }
 }
